+++ tests_unit/metrics/test_query_strings.py
